fix tests

+++ flyteplugins/go/tasks/pluginmachinery/webapi/pluginconfig_flags_test.go
@@ -105,7 +105,7 @@ func TestPluginConfig_SetFlags(t *testing.T) {
 			testValue := "1"
 
 			cmdFlags.Set("readRateLimiter.qps", testValue)
-			if vInt, err := cmdFlags.GetInt("readRateLimiter.qps"); err == nil {
+			if vInt, err := cmdFlags.GetFloat64("readRateLimiter.qps"); err == nil {
 				testDecodeJson_PluginConfig(t, fmt.Sprintf("%v", vInt), &actual.ReadRateLimiter.QPS)
 
 			} else {
